+++ test/src/widgets/reactive_dropdown_field_test.dart
@@ -1,0 +1,217 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:reactive_forms/reactive_forms.dart';
+
+import 'reactive_dropdown_testing_widget.dart';
+
+void main() {
+  group('ReactiveDropdownField Tests', () {
+    testWidgets(
+      'Dropdown initialize empty if control default value is null',
+      (WidgetTester tester) async {
+        // Given: a form with and control with default value
+        final form = FormGroup({
+          'dropdown': FormControl<bool>(),
+        });
+
+        // And: a widget that is bind to the form
+        await tester.pumpWidget(ReactiveDropdownTestingWidget(form: form));
+
+        // When: gets dropdown
+        Type dropdownType =
+            DropdownButton<bool>(items: null, onChanged: null).runtimeType;
+        DropdownButton dropdown = tester.firstWidget(find.byType(dropdownType));
+
+        // Expect: dropdown value is null
+        expect(dropdown.value, null);
+      },
+    );
+
+    testWidgets(
+      'Dropdown initialize true if control default value is true',
+      (WidgetTester tester) async {
+        // Given: a form with and control with default value
+        final form = FormGroup({
+          'dropdown': FormControl<bool>(defaultValue: true),
+        });
+
+        // And: a widget that is bind to the form
+        await tester.pumpWidget(ReactiveDropdownTestingWidget(form: form));
+
+        // When: gets dropdown
+        Type dropdownType =
+            DropdownButton<bool>(items: null, onChanged: null).runtimeType;
+        DropdownButton dropdown = tester.firstWidget(find.byType(dropdownType));
+
+        // Expect: dropdown value is null
+        expect(dropdown.value, true);
+      },
+    );
+
+    testWidgets(
+      'Dropdown initialize false if control default value is false',
+      (WidgetTester tester) async {
+        // Given: a form with and control with default value
+        final form = FormGroup({
+          'dropdown': FormControl<bool>(defaultValue: false),
+        });
+
+        // And: a widget that is bind to the form
+        await tester.pumpWidget(ReactiveDropdownTestingWidget(form: form));
+
+        // When: gets dropdown
+        Type dropdownType =
+            DropdownButton<bool>(items: null, onChanged: null).runtimeType;
+        DropdownButton dropdown = tester.firstWidget(find.byType(dropdownType));
+
+        // Expect: dropdown value is null
+        expect(dropdown.value, false);
+      },
+    );
+
+    testWidgets(
+      'Dropdown changes value to true if control changes value to true',
+      (WidgetTester tester) async {
+        // Given: a form with and control with default value
+        final form = FormGroup({
+          'dropdown': FormControl<bool>(defaultValue: false),
+        });
+
+        // And: a widget that is bind to the form
+        await tester.pumpWidget(ReactiveDropdownTestingWidget(form: form));
+
+        // When: changes control value
+        form.formControl('dropdown').value = true;
+        await tester.pump();
+
+        // Then: dropdown value is equals to control
+        Type dropdownType =
+            DropdownButton<bool>(items: null, onChanged: null).runtimeType;
+        DropdownButton dropdown = tester.firstWidget(find.byType(dropdownType));
+
+        expect(dropdown.value, form.formControl('dropdown').value);
+      },
+    );
+
+    testWidgets(
+      'Dropdown changes value to false if control changes value to false',
+      (WidgetTester tester) async {
+        // Given: a form with and control with default value
+        final form = FormGroup({
+          'dropdown': FormControl<bool>(defaultValue: true),
+        });
+
+        // And: a widget that is bind to the form
+        await tester.pumpWidget(ReactiveDropdownTestingWidget(form: form));
+
+        // When: changes control value
+        form.formControl('dropdown').value = false;
+        await tester.pump();
+
+        // Then: dropdown value is equals to control
+        Type dropdownType =
+            DropdownButton<bool>(items: null, onChanged: null).runtimeType;
+        DropdownButton dropdown = tester.firstWidget(find.byType(dropdownType));
+
+        expect(dropdown.value, form.formControl('dropdown').value);
+      },
+    );
+
+    testWidgets(
+      'Assert error if items in null',
+      (WidgetTester tester) async {
+        // Given: a form with and control with default value
+        final dropdown = () => ReactiveDropdownField(
+              formControlName: 'someName',
+              items: null,
+            );
+
+        expect(dropdown, throwsAssertionError);
+      },
+    );
+
+    testWidgets(
+      'Assert error if items is empty list',
+      (WidgetTester tester) async {
+        // Given: a form with and control with default value
+        final dropdown = () => ReactiveDropdownField(
+              formControlName: 'someName',
+              items: [],
+            );
+
+        expect(dropdown, throwsAssertionError);
+      },
+    );
+
+    testWidgets(
+      'Assert Error if attributes in null',
+      (WidgetTester tester) async {
+        // decoration
+        expect(
+          () => ReactiveDropdownField<bool>(
+            formControlName: 'dropdown',
+            items: [DropdownMenuItem(value: true, child: Text(''))],
+            decoration: null,
+          ),
+          throwsAssertionError,
+          reason: 'decoration != null',
+        );
+
+        // elevation
+        expect(
+          () => ReactiveDropdownField<bool>(
+            formControlName: 'dropdown',
+            items: [DropdownMenuItem(value: true, child: Text(''))],
+            elevation: null,
+          ),
+          throwsAssertionError,
+          reason: 'elevation != null',
+        );
+
+        // iconSize
+        expect(
+          () => ReactiveDropdownField<bool>(
+            formControlName: 'dropdown',
+            items: [DropdownMenuItem(value: true, child: Text(''))],
+            iconSize: null,
+          ),
+          throwsAssertionError,
+          reason: 'iconSize != null',
+        );
+
+        // isDense
+        expect(
+          () => ReactiveDropdownField<bool>(
+            formControlName: 'dropdown',
+            items: [DropdownMenuItem(value: true, child: Text(''))],
+            isDense: null,
+          ),
+          throwsAssertionError,
+          reason: 'isDense != null',
+        );
+
+        // isExpanded
+        expect(
+          () => ReactiveDropdownField<bool>(
+            formControlName: 'dropdown',
+            items: [DropdownMenuItem(value: true, child: Text(''))],
+            isExpanded: null,
+          ),
+          throwsAssertionError,
+          reason: 'isExpanded != null',
+        );
+
+        // itemHeight
+        expect(
+          () => ReactiveDropdownField<bool>(
+            formControlName: 'dropdown',
+            items: [DropdownMenuItem(value: true, child: Text(''))],
+            itemHeight: 0,
+          ),
+          throwsAssertionError,
+          reason: 'itemHeight == null || itemHeight > 0',
+        );
+      },
+    );
+  });
+}